correcting bugs from light mode to dark mode
Theres a bug that the app could't change the modes from dark to light, and the switch was on even with the whole app in dark mode. However the problem was in the file theme_provider.dart that was verifying if the darkmode is equal to the mode that the app was using, making a mess in the settingspage when i used them.